requirements.txt
# Game Asset Pipeline - Dependencies
Pillow>=10.0.0
numpy>=1.24.0
numpy-stl>=3.0.0
trimesh>=4.0.0
Flask>=3.0.0
PyYAML>=6.0
matplotlib>=3.7.0
Jinja2>=3.1.0
